Refactor some files for custom responses file.
