+++ dces-drc-integration/src/main/java/uk/gov/justice/laa/crime/dces/integration/model/exception/trace/CustomDataResponse.java
@@ -1,4 +1,4 @@
-package uk.gov.justice.laa.crime.dces.integration.model.exception.context;
+package uk.gov.justice.laa.crime.dces.integration.model.exception.trace;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class ErrorResponse<T> {
+public class CustomDataResponse<T> {
     private T data;
     private TraceData trace;
 }
