+++ dces-drc-integration/src/main/java/uk/gov/justice/laa/crime/dces/integration/model/exception/trace/WrapResponse.java
@@ -1,4 +1,4 @@
-package uk.gov.justice.laa.crime.dces.integration.model.exception.context;
+package uk.gov.justice.laa.crime.dces.integration.model.exception.trace;
 
 
 import java.lang.annotation.ElementType;
